+++ src/utils/err_codes.py
@@ -8,9 +8,10 @@ def iota(set_val: int | None = None):
     return tmp
 
 
+ERR_UNK_ERR = iota(-1)
 ERR_ALL_GOOD = iota(0)
 ERR_FALSE = iota()
-# MAIN PROG
+# MAIN PROGRAM
 ERR_MP_EXPD_VAL_OPT = iota(2)
 ERR_MP_UNK_TOK = iota()
 ERR_MP_INV_VAL = iota()